Fikset valideringsmetode digitsOnly slik at den ikke feiler ved stoore tall. Endret fra Integer til Long.
git-svn-id: http://projects.knowit.no/repos/frikomport/branches/SVV@916 b0a0c851-5f35-0410-b037-9bcbcc8e1326

+++ src/web/no/unified/soak/webapp/action/BaseFormController.java
@@ -517,7 +517,7 @@ public class BaseFormController extends SimpleFormController {
 						return nErrors;
 					}
 					
-					Integer methodResultInt = Integer.parseInt(methodResultStr);
+					Long methodResultInt = Long.parseLong(methodResultStr);
 					if (methodResultInt < 0) {
 						String fieldText = getFieldDisplayName(obj, referencingObjectFieldnamePrefix+fieldName);
 						Object[] args = new Object[] { fieldText };
